requirements.txt
cachetools==5.3.0
google-auth==2.16.3
google-auth-oauthlib==1.0.0
gspread==5.7.2
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2023.2
requests-oauthlib==1.3.1
rsa==4.9
termcolor==2.2.0
